ssh keys: parse keys on demand
This demonstrates an idea we tossed around a while back; save response
data on a type and dynamically parse the data as needed.

To demonstrate one approach I've modified the `ssh_keys` type to save
the json response data and include a function pointer, `next()`, to
dynamically parse and return the next available ssh key.